+++ src/main/java/org/dimdev/dimdoors/util/Weighted.java
@@ -5,5 +5,5 @@ public interface Weighted<P> {
 	Should always return the same number if the same parameters are provided.
 	returned number should always be >= 0
 	 */
-	int getWeight(P parameters);
+	double getWeight(P parameters);
 }
